+++ packages/sync-server/app/tasks/CertificateNotificationProcessor.js
@@ -15,7 +15,6 @@ import { getLocalisation } from '../localisation';
 import {
   createProofOfVaccination,
   VdsNcDocument,
-  vdsConfig,
   generateUVCIForPatient,
 } from '../integrations/VdsNc';
 
@@ -34,7 +33,7 @@ export class CertificateNotificationProcessor extends ScheduledTask {
   async run() {
     const { models } = this.context.store;
     const { CertificateNotification, PatientCommunication, Patient } = models;
-    const vdsEnabled = vdsConfig().enabled;
+    const vdsEnabled = config.integrations.vdsNc.enabled;
     const localisation = await getLocalisation();
 
     const queuedNotifications = await CertificateNotification.findAll({